requirements.txt
python-telegram-bot==20.0
requests==2.26.0
pillow==8.4.0
Flask==2.0.1